refactor: 인터페이스 파일명 수정
- .d 제거
- 파일명을 인터페이스 명과 일치하게 변경
